+++ examples/script_02.py
@@ -9,7 +9,7 @@
 import re
 import sys
 from pathlib import Path
-from subprocess import (
+from subprocess import (  # ruff: ignore[suspicious-subprocess-import]
     check_output,
 )