+++ incolume/py/githooks/core/__init__.py
@@ -3,6 +3,7 @@
 from incolume.py.githooks.core.__main__ import (
     __package_name__,
     __version__,
+    backup_file,
     debug_enable,
     debug_var_active,
     get_branchname,
@@ -19,6 +20,7 @@ __all__ = [
     'CONTEXT_SETTINGS_CLICK',
     '__package_name__',
     '__version__',
+    'backup_file',
     'debug_enable',
     'debug_var_active',
     'get_branchname',